Renamed stream-src -> stream_src and made it a package
