+++ stream_src/122012B1.DBF
