+++ stream_src/common.py
